Module Decimal__.Calc

val z2 : Z.t

val z10 : Z.t

val z100 : Z.t

val zeros : Str.regexp

val unreliable : Str.regexp

Unreliable digits at the end of the log10_digits calculation

val decimal_lshift_exact : Z.t -> int -> Z.t option

decimal_lshift_exact n e is Some (n * 10 ** e) if it's an integer, else None.

val sqrt_nearest : Z.t -> Z.t -> Z.t -> Z.t

val sqrt_nearest : Z.t -> Z.t -> Z.t

sqrt_nearest n a is the closest integer to the square root of the positive integer n. a is an initial approximation of the square root. Any positive integer will do for a, but the closer a is to the square root of n the faster convergence will be.

val rshift_nearest : Z.t -> int -> Z.t

rshift_nearest x shift is the closest integer to x / 2**shift, where shift is non-negative. Uses round-to-even in case of a tie.

val div_nearest : Z.t -> Z.t -> Z.t

val ilog : ?⁠l:int -> Z.t -> Z.t -> Z.t

ilog ?l x m is the integer approximation to m * log (x / m), with absolute error boundable in terms only of x / m.

Given positive integers x and m, return an integer approximation to m * log (x / m). For l = 8 and 0.1 <= x / m <= 10 the difference between the approximation and the exact result is at most 22. For l = 8 and 1.0 <= x / m <= 10.0 the difference is at most 15. In both cases these are upper bounds on the error; it will usually be much smaller.

val log10_digits : string Stdlib.ref

val log10_digits : int -> Z.t

log10_digits p is (floor 10**p) * log 10, given p >= 0. For example, log10_digits 3 is 2302.

val dlog10 : Z.t -> int -> int -> Z.t

dlog10 c e p is an integer approximation of 10**p * log10 (c * 10**e), with an absolute error of at most 1. Assumes that:

c > 0

p >= 0

c * 10**e is not exactly 1

val dlog : Z.t -> int -> int -> Z.t

dlog c e p is an integer approximation of 10**p * log (c * 10 * e), with an absolute error of at most 1. Assumes that c * 10 * e is not exactly 1.

val iexp : ?⁠l:int -> Z.t -> Z.t -> Z.t

iexp ?l x m is an integer approximation of m * exp (x / m), given m > 0 and such that x / m is small in absolute value. For 0 <= x / m <= 2.4, the absolute error in the result is bounded by 60 (and is usually much smaller).

val dexp : Z.t -> int -> int -> Z.t * int

dexp c e p is an approximation to exp (c * 10**e), with p decimal places of precision.

Returns integers d, f such that:

10**(p - 1) <= d <= 10**p, and

(d - 1) * 10**f < exp (c * 10**e) < (d + 1) * 10**f

In other words, d * 10**f is an approximation to exp (c * 10**e) with p digits of precision, and with an error in d of at most 1. This is almost, but not quite, the same as the error being < 1ulp: when d = 10**(p - 1) the error could be up to 10 ulp.

val dpower : Z.t -> int -> Z.t -> int -> int -> Z.t * int

dpower xc xe yc ye p is x ** y, given integers xc, xe, yc, and ye representing decimals x = xc * 10**xe and y = yc * 10**ye. Returns a pair of integers c, e such that:

10**(p - 1) <= c <= 10**p, and

(c - 1) * 10**e < x**y < (c + 1) * 10**e

In other words, c * 10**e is an approximation to x**y with p digits of precision, and with an error in c of at most 1. This almost, but not quite, the same as the error being < 1ulp: when c = 10**(p - 1) we can only guarantee error < 10ulp.

We assume that: x is positive and not equal to 1, and y is nonzero.

val log10_lb : ?⁠correction:(char * int) list -> Z.t -> int

log10_lb ?correction c is a lower bound for 100 * log10 c for a positive integer c.