Module Decimal.Context

Settings that control precision, rounding mode, exceptional behaviour, etc.

type round =

| Down
Round towards 0; truncate.

| Up
Round away from 0.

| Half_up
Round up if last significant digit is >= 5, else round down.

| Half_down
Round up if last significant digit is > 5, else round down.

| Half_even
Round up if last significant digit is > 5 or next-to-last significant digit is odd, else round down.

| Ceiling
Round up if last significant digit is > 0, else no change.

| Floor
Round down if last significant digit is > 0, else no change.

| Zero_five_up
Round zero or five away from 0.

type t

Controls, precision, rounding, traps (exception handling), etc. settings.

val default : unit -> t

default () is the current thread-local context. May change between calls.

val set_default : t -> unit

set_default t sets t as the current thread-local context.

val make : ?⁠prec:int -> ?⁠round:round -> ?⁠e_max:int -> ?⁠e_min:int -> ?⁠capitals:bool -> ?⁠clamp:bool -> unit -> t

make ?prec ?round ?e_max ?e_min ?capitals ?clamp () is a new context value with the given settings and the following traps configured:

conversion_syntax

invalid_operation

div_by_zero

overflow

These may be overridden by using Signal.set (Context.traps context) id bool, or setting a new default context.

val copy : orig:t -> ?⁠prec:int -> ?⁠round:round -> ?⁠e_max:int -> ?⁠e_min:int -> ?⁠capitals:bool -> ?⁠clamp:bool -> ?⁠traps:Signal.array -> ?⁠flags:Signal.array -> unit -> t

copy ~orig ?prec ?round ?e_max ?e_min ?capitals ?clamp ?traps ?flags is a deep copy of orig with the given settings. The optional settings default to the same values as those in orig.

val prec : t -> int

val round : t -> round

val e_max : t -> int

val e_min : t -> int

val capitals : t -> bool

val clamp : t -> bool

val traps : t -> Signal.array

val flags : t -> Signal.array

val e_tiny : t -> int

e_tiny t is e_min t - prec t + 1, the minimum allowable exponent of context t.

val e_top : t -> int

e_top t is the maximum exponent of context t.