Module Decimal__.Calc

val z2 : Z.t

val z10 : Z.t

val z100 : Z.t

val zeros : Str.regexp

val unreliable : Str.regexp

Unreliable digits at the end of the log10_digits calculation

val decimal_lshift_exact : Z.t -> int -> Z.t option

decimal_lshift_exact n e is Some (n * 10 ** e) if it's an integer, else None.

val sqrt_nearest : Z.t -> Z.t -> Z.t -> Z.t

val sqrt_nearest : Z.t -> Z.t -> Z.t

sqrt_nearest n a is the closest integer to the square root of the positive integer n. a is an initial approximation of the square root. Any positive integer will do for a, but the closer a is to the square root of n the faster convergence will be.

val rshift_nearest : Z.t -> int -> Z.t

rshift_nearest x shift is the closest integer to x / 2**shift, where shift is non-negative. Uses round-to-even in case of a tie.

val div_nearest : Z.t -> Z.t -> Z.t

val ilog : ?⁠l:int -> Z.t -> Z.t -> Z.t

ilog ?l x m is the integer approximation to m * log (x / m), with absolute error boundable in terms only of x / m.

Given positive integers x and m, return an integer approximation to m * log (x / m). For l = 8 and 0.1 <= x / m <= 10 the difference between the approximation and the exact result is at most 22. For l = 8 and 1.0 <= x / m <= 10.0 the difference is at most 15. In both cases these are upper bounds on the error; it will usually be much smaller.

val log10_digits : string Stdlib.ref

val log10_digits : int -> Z.t

log10_digits p is (floor 10**p) * log 10, given p >= 0. For example, log10_digits 3 is 2302.

val dlog10 : Z.t -> int -> int -> Z.t

dlog10 c e p is an integer approximation of 10**p * log10 (c * 10**e), with an absolute error of at most 1. Assumes that:

c > 0

p >= 0

c * 10**e is not exactly 1